+++ src/git.hpp
@@ -1,0 +1,12 @@
+#ifndef GIT_TUI_GIT_HPP
+#define GIT_TUI_GIT_HPP
+
+namespace gittui::git {
+int main(int argc, const char** argv);
+}  // namespace gittui::git
+
+#endif /* end of include guard: GIT_TUI_GIT_HPP */
+
+// Copyright 2024 Arthur Sonzogni. All rights reserved.
+// Use of this source code is governed by the MIT license that can be found in
+// the LICENSE file.