+++ src/main.cpp
@@ -1,6 +1,7 @@
-#include <string>  // for operator==, allocator, basic_string, string
+#include <string>  // for operator==, basic_string, allocator, string
 
 #include "diff.hpp"     // for main
+#include "git.hpp"      // for main
 #include "help.hpp"     // for main
 #include "log.hpp"      // for main
 #include "version.hpp"  // for main
@@ -28,10 +29,20 @@ int main(int argc, const char** argv) {
   if (command == "log")
     return gittui::log::main(argc, argv);
 
-  if (command == "version")
+  if (command == "version" || command == "--version" || command == "-v")
     return gittui::version::main(argc, argv);
 
-  return gittui::help::main(argc, argv);
+  if (command == "help" || command == "--help" || command == "-h")
+    return gittui::help::main(argc, argv);
+
+  if (command == "git") {
+    return gittui::git::main(argc, argv);
+  }
+
+  // Unknown command, fallback to git, uneat the first argument.
+  argc++;
+  argv--;
+  return gittui::git::main(argc, argv);
 }
 
 // Copyright 2021 Arthur Sonzogni. All rights reserved.